Addded core structure for the simulation

+++ SimProject/SimProject/includes/timer_manager.h
@@ -15,13 +15,12 @@ namespace TimerManager
 	/// and current frame to time to delta time
 	/// </summary>
 	void Update();
-
 	/// <summary>
 	/// Set a delay to infinetly call a function with a Delay should be > 0
 	/// </summary>
 	/// <param name="function -->">The function to call when timer expires</param>
 	/// <param name="delay -->">Delay in seconds before calling the function</param>    
-	void SetTimerByEvent(void function(),int delay,float deltaTime); 
+	void SetTimerByEvent(void function(),float delay,float deltaTime); 
 }
